+++ src/react/src/components/Contribute/ProductionLocationInfo.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useMemo, useState, useRef } from 'react';
+import React, { useEffect, useState, useRef } from 'react';
 import { useLocation, useParams, useHistory } from 'react-router-dom';
 import { array, bool, func, number, object, string } from 'prop-types';
 import { connect } from 'react-redux';
@@ -36,7 +36,10 @@ import {
     DISABLE_LIST_UPLOADING,
     MAINTENANCE_MESSAGE,
 } from '../../util/constants';
-import { useResetScrollPosition } from '../../util/hooks';
+import {
+    useResetScrollPosition,
+    useSingleLocationContributionForm,
+} from '../../util/hooks';
 import {
     countryOptionsPropType,
     facilityProcessingTypeOptionsPropType,
@@ -48,10 +51,7 @@ import {
     mapDjangoChoiceTuplesToSelectOptions,
     mapFacilityTypeOptions,
     mapProcessingTypeOptions,
-    isValidNumberOfWorkers,
-    isRequiredFieldValid,
     getSelectStyles,
-    getNumberOfWorkersValidationError,
 } from '../../util/util';
 
 import FeatureFlag from '../FeatureFlag';
@@ -62,6 +62,7 @@ import StyledTooltip from '../StyledTooltip';
 
 import InputErrorText from './InputErrorText';
 import ProductionLocationDialog from './ProductionLocationDialog';
+import PostContributionSubmitErrorNotification from './PostContributionSubmitErrorNotification/PostContributionSubmitErrorNotification';
 
 const ProductionLocationInfo = ({
     submitMethod,
@@ -95,93 +96,21 @@ const ProductionLocationInfo = ({
     const history = useHistory();
     const { moderationID, osID } = useParams();
 
-    const queryParams = new URLSearchParams(location.search);
-    const nameInQuery = queryParams.get('name');
-    const addressInQuery = queryParams.get('address');
-    const countryInQuery = queryParams.get('country');
-    const [inputName, setInputName] = useState(nameInQuery ?? '');
-    const [inputAddress, setInputAddress] = useState(addressInQuery ?? '');
-    const [inputCountry, setInputCountry] = useState(null);
-    const [nameTouched, setNameTouched] = useState(false);
-    const [addressTouched, setAddressTouched] = useState(false);
-    const [countryTouched, setCountryTouched] = useState(false);
-    const [enabledTaxonomy, setEnabledTaxonomy] = useState(false);
-    const [sector, setSector] = useState('');
-    const [productType, setProductType] = useState([]);
-    const [locationType, setLocationType] = useState(null);
-    const [processingType, setProcessingType] = useState(null);
-    const [numberOfWorkers, setNumberOfWorkers] = useState('');
-    const [parentCompany, setParentCompany] = useState('');
     const customSelectComponents = { DropdownIndicator: null };
-    const isCountryError = countryTouched && !inputCountry?.value;
-
-    const resetAdditionalDataFields = () => {
-        setSector('');
-        setProductType([]);
-        setLocationType(null);
-        setProcessingType(null);
-        setNumberOfWorkers('');
-        setParentCompany('');
-    };
 
     useResetScrollPosition(location);
-
-    const inputData = useMemo(
-        () => ({
-            name: inputName,
-            address: inputAddress,
-            country: inputCountry,
-            sector,
-            productType,
-            locationType,
-            processingType,
-            numberOfWorkers,
-            parentCompany,
-        }),
-        [
-            inputName,
-            inputAddress,
-            inputCountry,
-            sector,
-            productType,
-            locationType,
-            processingType,
-            numberOfWorkers,
-            parentCompany,
-        ],
-    );
 
     const [
         showProductionLocationDialog,
         setShowProductionLocationDialog,
     ] = useState(null);
 
-    const handleNameChange = event => {
-        setInputName(event.target.value);
-    };
-    const handleAddressChange = event => {
-        setInputAddress(event.target.value);
-    };
-    const handleParentCompanyChange = event => {
-        setParentCompany(event.target.value);
-    };
+    const [
+        showPostSubmitErrorNotification,
+        setShowPostSubmitErrorNotification,
+    ] = useState(false);
 
-    const handleNameBlur = () => {
-        setNameTouched(true);
-    };
-    const handleAddressBlur = () => {
-        setAddressTouched(true);
-    };
-    const handleCountryBlur = () => {
-        setCountryTouched(true);
-    };
-
-    const isFormValid = !!(
-        isRequiredFieldValid(inputName) &&
-        isRequiredFieldValid(inputAddress) &&
-        inputCountry?.value &&
-        isValidNumberOfWorkers(numberOfWorkers)
-    );
+    const [enabledTaxonomy, setEnabledTaxonomy] = useState(false);
 
     let handleProductionLocation;
     switch (submitMethod) {
@@ -199,11 +128,28 @@ const ProductionLocationInfo = ({
             break;
     }
 
+    const contributionForm = useSingleLocationContributionForm(values => {
+        setShowPostSubmitErrorNotification(false);
+        handleProductionLocation(values);
+    });
+
     const instructionExtraMessage =
         submitMethod === 'PATCH'
             ? 'These fields are pre-filled with the data from your search, but you can edit them.'
             : '';
     const submitButtonText = submitMethod === 'POST' ? 'Submit' : 'Update';
+
+    const resetAdditionalDataFields = () => {
+        contributionForm.setValues({
+            ...contributionForm.values,
+            sector: [],
+            productType: [],
+            locationType: [],
+            processingType: [],
+            numberOfWorkers: '',
+            parentCompany: '',
+        });
+    };
 
     const [showAdditionalInfo, setShowAdditionalInfo] = useState(false);
     const onSwitchChange = () => {
@@ -242,14 +188,25 @@ const ProductionLocationInfo = ({
 
     useEffect(() => {
         if (singleProductionLocationData && osID) {
-            setInputName(singleProductionLocationData.name ?? '');
-            setInputAddress(singleProductionLocationData.address ?? '');
-            if (singleProductionLocationData.country) {
-                setInputCountry({
-                    value: singleProductionLocationData?.country.alpha_2,
-                    label: singleProductionLocationData?.country.name,
-                });
-            }
+            contributionForm.setValues({
+                ...contributionForm.values,
+                name: singleProductionLocationData.name ?? '',
+                address: singleProductionLocationData.address ?? '',
+                country: singleProductionLocationData.country
+                    ? {
+                          value: singleProductionLocationData?.country.alpha_2,
+                          label: singleProductionLocationData?.country.name,
+                      }
+                    : null,
+            });
+            contributionForm.setTouched(
+                {
+                    name: true,
+                    address: true,
+                    country: true,
+                },
+                false,
+            );
         }
     }, [singleProductionLocationData, osID]);
 
@@ -276,15 +233,6 @@ const ProductionLocationInfo = ({
             fetchCountries();
         }
     }, [countriesOptions, fetchCountries]);
-
-    useEffect(() => {
-        if (countriesOptions && !isEmpty(countryInQuery)) {
-            const prefilledCountry = countriesOptions.filter(
-                el => el.value === countryInQuery,
-            );
-            setInputCountry(prefilledCountry[0]);
-        }
-    }, [countriesOptions]);
 
     useEffect(() => {
         if (!facilityProcessingTypeOptions) {
@@ -340,7 +288,7 @@ const ProductionLocationInfo = ({
 
     useEffect(() => {
         if (!pendingModerationEventFetching && pendingModerationEventError) {
-            toast(toString(pendingModerationEventError));
+            setShowPostSubmitErrorNotification(true);
         }
     }, [pendingModerationEventFetching, pendingModerationEventError]);
 
@@ -375,9 +323,10 @@ const ProductionLocationInfo = ({
 
     useEffect(() => {
         setEnabledTaxonomy(
-            sector.length === 1 && sector[0].value === 'Apparel',
+            contributionForm.values.sector.length === 1 &&
+                contributionForm.values.sector[0].value === 'Apparel',
         );
-    }, [sector]);
+    }, [contributionForm.values.sector]);
 
     useEffect(
         () => () => {
@@ -394,13 +343,31 @@ const ProductionLocationInfo = ({
         }
     }, [singleProductionLocationError]);
 
+    /*
+    The Formik library, which is used for the SLC form, doesn’t
+    automatically re-run validation when the value of the product
+    type and country field changes - unless the field is actively touched or
+    blurred. This issue happens only when clicking the "x"
+    (remove) button on a value inside the React Select
+    multi-select field.
+    A similar issue occurs with the country field - after setting
+    the value, validation isn’t triggered to avoid showing an
+    inappropriate error.
+    Because of this, validation is manually triggered using
+    useEffect and validateField.
+    */
+    useEffect(() => {
+        contributionForm.validateField('productType');
+        contributionForm.validateField('country');
+    }, [contributionForm.values.productType, contributionForm.values.country]);
+
     const activeSubmitButton = (
         <Button
             color="secondary"
             variant="contained"
-            onClick={() => handleProductionLocation(inputData, osID)}
+            onClick={contributionForm.handleSubmit}
             className={classes.submitButtonStyles}
-            disabled={!isFormValid}
+            disabled={!contributionForm.isValid}
         >
             {submitButtonText}
         </Button>
@@ -472,9 +439,19 @@ const ProductionLocationInfo = ({
                         <TextField
                             id="name"
                             className={classes.textInputStyles}
-                            value={inputName}
-                            onChange={handleNameChange}
-                            onBlur={handleNameBlur}
+                            value={contributionForm.values.name}
+                            onChange={e => {
+                                contributionForm.setFieldValue(
+                                    'name',
+                                    e.target.value,
+                                );
+                                contributionForm.setFieldTouched(
+                                    'name',
+                                    true,
+                                    false,
+                                );
+                            }}
+                            onBlur={contributionForm.handleBlur}
                             placeholder="Enter the name"
                             variant="outlined"
                             aria-label="Enter the name"
@@ -482,8 +459,8 @@ const ProductionLocationInfo = ({
                                 classes: {
                                     input: `
                                     ${
-                                        nameTouched &&
-                                        !isRequiredFieldValid(inputName) &&
+                                        contributionForm.touched.name &&
+                                        contributionForm.errors.name &&
                                         classes.errorStyle
                                     }`,
                                     notchedOutline:
@@ -491,16 +468,19 @@ const ProductionLocationInfo = ({
                                 },
                             }}
                             helperText={
-                                nameTouched &&
-                                !isRequiredFieldValid(inputName) && (
-                                    <InputErrorText />
+                                contributionForm.touched.name &&
+                                contributionForm.errors.name && (
+                                    <InputErrorText
+                                        text={contributionForm.errors.name}
+                                    />
                                 )
                             }
                             FormHelperTextProps={{
                                 className: classes.helperText,
                             }}
                             error={
-                                nameTouched && !isRequiredFieldValid(inputName)
+                                contributionForm.touched.name &&
+                                !!contributionForm.errors.name
                             }
                         />
                     </div>
@@ -523,9 +503,19 @@ const ProductionLocationInfo = ({
                         <TextField
                             id="address"
                             className={classes.textInputStyles}
-                            value={inputAddress}
-                            onChange={handleAddressChange}
-                            onBlur={handleAddressBlur}
+                            value={contributionForm.values.address}
+                            onChange={e => {
+                                contributionForm.setFieldValue(
+                                    'address',
+                                    e.target.value,
+                                );
+                                contributionForm.setFieldTouched(
+                                    'address',
+                                    true,
+                                    false,
+                                );
+                            }}
+                            onBlur={contributionForm.handleBlur}
                             placeholder="Enter the full address"
                             variant="outlined"
                             aria-label="Enter the address"
@@ -533,8 +523,8 @@ const ProductionLocationInfo = ({
                                 classes: {
                                     input: `${classes.searchInputStyles}
                                 ${
-                                    addressTouched &&
-                                    !isRequiredFieldValid(inputAddress) &&
+                                    contributionForm.touched.address &&
+                                    contributionForm.errors.address &&
                                     classes.errorStyle
                                 }`,
                                     notchedOutline:
@@ -542,17 +532,19 @@ const ProductionLocationInfo = ({
                                 },
                             }}
                             helperText={
-                                addressTouched &&
-                                !isRequiredFieldValid(inputAddress) && (
-                                    <InputErrorText />
+                                contributionForm.touched.address &&
+                                contributionForm.errors.address && (
+                                    <InputErrorText
+                                        text={contributionForm.errors.address}
+                                    />
                                 )
                             }
                             FormHelperTextProps={{
                                 className: classes.helperText,
                             }}
                             error={
-                                addressTouched &&
-                                !isRequiredFieldValid(inputAddress)
+                                contributionForm.touched.address &&
+                                !!contributionForm.errors.address
                             }
                         />
                     </div>
@@ -577,19 +569,40 @@ const ProductionLocationInfo = ({
                             name="country"
                             aria-label="Country"
                             options={countriesOptions || []}
-                            value={inputCountry}
-                            onChange={setInputCountry}
-                            onBlur={handleCountryBlur}
+                            value={contributionForm.values.country}
+                            onChange={value => {
+                                contributionForm.setFieldValue(
+                                    'country',
+                                    value,
+                                );
+                                contributionForm.setFieldTouched(
+                                    'country',
+                                    true,
+                                    false,
+                                );
+                            }}
+                            onBlur={() =>
+                                contributionForm.setFieldTouched(
+                                    'country',
+                                    true,
+                                )
+                            }
                             className={classes.selectStyles}
-                            styles={getSelectStyles(isCountryError)}
+                            styles={getSelectStyles(
+                                contributionForm.touched.country &&
+                                    !!contributionForm.errors.country,
+                            )}
                             placeholder="Country"
                             isMulti={false}
                         />
-                        {isCountryError && (
-                            <div className={classes.errorWrapStyles}>
-                                <InputErrorText />
-                            </div>
-                        )}
+                        {contributionForm.touched.country &&
+                            contributionForm.errors.country && (
+                                <div className={classes.errorWrapStyles}>
+                                    <InputErrorText
+                                        text={contributionForm.errors.country}
+                                    />
+                                </div>
+                            )}
                     </div>
                     <hr className={classes.separator} />
                     <div
@@ -650,8 +663,13 @@ const ProductionLocationInfo = ({
                                                 mockedSectors,
                                             ) || []
                                         }
-                                        value={sector}
-                                        onChange={setSector}
+                                        value={contributionForm.values.sector}
+                                        onChange={value =>
+                                            contributionForm.setFieldValue(
+                                                'sector',
+                                                value,
+                                            )
+                                        }
                                         styles={getSelectStyles()}
                                         className={classes.selectStyles}
                                         placeholder="Select sector(s)"
@@ -678,14 +696,46 @@ const ProductionLocationInfo = ({
                                     <StyledSelect
                                         creatable
                                         name="product-type"
-                                        value={productType}
-                                        onChange={setProductType}
+                                        value={
+                                            contributionForm.values.productType
+                                        }
+                                        onChange={value => {
+                                            contributionForm.setFieldValue(
+                                                'productType',
+                                                value,
+                                            );
+                                            contributionForm.setFieldTouched(
+                                                'productType',
+                                                true,
+                                                false,
+                                            );
+                                        }}
                                         placeholder="Enter product type(s)"
                                         aria-label="Enter product type(s)"
-                                        styles={getSelectStyles()}
+                                        styles={getSelectStyles(
+                                            contributionForm.touched
+                                                .productType &&
+                                                !!contributionForm.errors
+                                                    .productType,
+                                        )}
                                         className={classes.selectStyles}
                                         components={customSelectComponents}
                                     />
+                                    {contributionForm.touched.productType &&
+                                        contributionForm.errors.productType && (
+                                            <div
+                                                className={
+                                                    classes.errorWrapStyles
+                                                }
+                                            >
+                                                <InputErrorText
+                                                    text={
+                                                        contributionForm.errors
+                                                            .productType
+                                                    }
+                                                />
+                                            </div>
+                                        )}
                                 </div>
                                 <div
                                     className={`${classes.inputSectionWrapStyles} ${classes.wrapStyles}`}
@@ -713,10 +763,19 @@ const ProductionLocationInfo = ({
                                             options={mapFacilityTypeOptions(
                                                 facilityProcessingTypeOptions ||
                                                     [],
-                                                processingType || [],
+                                                contributionForm.values
+                                                    .processingType,
                                             )}
-                                            value={locationType}
-                                            onChange={setLocationType}
+                                            value={
+                                                contributionForm.values
+                                                    .locationType
+                                            }
+                                            onChange={value =>
+                                                contributionForm.setFieldValue(
+                                                    'locationType',
+                                                    value,
+                                                )
+                                            }
                                             styles={getSelectStyles()}
                                             className={classes.selectStyles}
                                             placeholder="Select location type(s)"
@@ -725,8 +784,16 @@ const ProductionLocationInfo = ({
                                         <StyledSelect
                                             creatable
                                             name="location-type"
-                                            value={locationType || []}
-                                            onChange={setLocationType}
+                                            value={
+                                                contributionForm.values
+                                                    .locationType
+                                            }
+                                            onChange={value =>
+                                                contributionForm.setFieldValue(
+                                                    'locationType',
+                                                    value,
+                                                )
+                                            }
                                             placeholder="Enter location type(s)"
                                             aria-label="Location type"
                                             styles={getSelectStyles()}
@@ -761,10 +828,19 @@ const ProductionLocationInfo = ({
                                             options={mapProcessingTypeOptions(
                                                 facilityProcessingTypeOptions ||
                                                     [],
-                                                locationType || [],
+                                                contributionForm.values
+                                                    .locationType,
                                             )}
-                                            value={processingType}
-                                            onChange={setProcessingType}
+                                            value={
+                                                contributionForm.values
+                                                    .processingType
+                                            }
+                                            onChange={value =>
+                                                contributionForm.setFieldValue(
+                                                    'processingType',
+                                                    value,
+                                                )
+                                            }
                                             styles={getSelectStyles()}
                                             className={classes.selectStyles}
                                             placeholder="Select processing type(s)"
@@ -773,8 +849,16 @@ const ProductionLocationInfo = ({
                                         <StyledSelect
                                             creatable
                                             name="processing-type"
-                                            value={processingType || []}
-                                            onChange={setProcessingType}
+                                            value={
+                                                contributionForm.values
+                                                    .processingType
+                                            }
+                                            onChange={value =>
+                                                contributionForm.setFieldValue(
+                                                    'processingType',
+                                                    value,
+                                                )
+                                            }
                                             placeholder="Enter processing type(s)"
                                             aria-label="Processing Type"
                                             styles={getSelectStyles()}
@@ -803,27 +887,39 @@ const ProductionLocationInfo = ({
                                     <TextField
                                         id="number_of_workers"
                                         error={
-                                            !isValidNumberOfWorkers(
-                                                numberOfWorkers,
-                                            )
+                                            contributionForm.touched
+                                                .numberOfWorkers &&
+                                            !!contributionForm.errors
+                                                .numberOfWorkers
                                         }
                                         variant="outlined"
                                         className={classes.textInputStyles}
-                                        value={numberOfWorkers}
-                                        onChange={e =>
-                                            setNumberOfWorkers(
-                                                e.target.value.trim(),
-                                            )
+                                        value={
+                                            contributionForm.values
+                                                .numberOfWorkers
                                         }
+                                        onChange={e => {
+                                            contributionForm.setFieldValue(
+                                                'numberOfWorkers',
+                                                e.target.value,
+                                            );
+                                            contributionForm.setFieldTouched(
+                                                'numberOfWorkers',
+                                                true,
+                                                false,
+                                            );
+                                        }}
                                         placeholder="Enter the number of workers as a number or range"
                                         helperText={
-                                            !isValidNumberOfWorkers(
-                                                numberOfWorkers,
-                                            ) && (
+                                            contributionForm.errors
+                                                .numberOfWorkers &&
+                                            contributionForm.touched
+                                                .numberOfWorkers && (
                                                 <InputErrorText
-                                                    text={getNumberOfWorkersValidationError(
-                                                        numberOfWorkers,
-                                                    )}
+                                                    text={
+                                                        contributionForm.errors
+                                                            .numberOfWorkers
+                                                    }
                                                 />
                                             )
                                         }
@@ -834,9 +930,11 @@ const ProductionLocationInfo = ({
                                             classes: {
                                                 input: `
                                             ${
-                                                !isValidNumberOfWorkers(
-                                                    numberOfWorkers,
-                                                ) && classes.errorStyle
+                                                contributionForm.errors
+                                                    .numberOfWorkers &&
+                                                contributionForm.touched
+                                                    .numberOfWorkers &&
+                                                classes.errorStyle
                                             }`,
                                                 notchedOutline:
                                                     classes.notchedOutlineStyles,
@@ -864,22 +962,72 @@ const ProductionLocationInfo = ({
                                     <TextField
                                         id="parent_company"
                                         className={classes.textInputStyles}
-                                        value={parentCompany}
-                                        onChange={handleParentCompanyChange}
+                                        value={
+                                            contributionForm.values
+                                                .parentCompany
+                                        }
+                                        onChange={e => {
+                                            contributionForm.setFieldValue(
+                                                'parentCompany',
+                                                e.target.value,
+                                            );
+                                            contributionForm.setFieldTouched(
+                                                'parentCompany',
+                                                true,
+                                                false,
+                                            );
+                                        }}
                                         placeholder="Enter the parent company"
                                         variant="outlined"
                                         aria-label="Parent company"
                                         InputProps={{
                                             classes: {
+                                                input: `${
+                                                    contributionForm.touched
+                                                        .parentCompany &&
+                                                    contributionForm.errors
+                                                        .parentCompany &&
+                                                    classes.errorStyle
+                                                }`,
                                                 notchedOutline:
                                                     classes.notchedOutlineStyles,
                                             },
                                         }}
+                                        helperText={
+                                            contributionForm.touched
+                                                .parentCompany &&
+                                            contributionForm.errors
+                                                .parentCompany && (
+                                                <InputErrorText
+                                                    text={
+                                                        contributionForm.errors
+                                                            .parentCompany
+                                                    }
+                                                />
+                                            )
+                                        }
+                                        FormHelperTextProps={{
+                                            className: classes.helperText,
+                                        }}
+                                        error={
+                                            contributionForm.touched
+                                                .parentCompany &&
+                                            !!contributionForm.errors
+                                                .parentCompany
+                                        }
                                     />
                                 </div>
                             </>
                         )}
                     </div>
+                    {showPostSubmitErrorNotification && (
+                        <PostContributionSubmitErrorNotification
+                            showNotification={
+                                setShowPostSubmitErrorNotification
+                            }
+                            errorObj={pendingModerationEventError}
+                        />
+                    )}
                     <div className={classes.buttonsContainerStyles}>
                         <Button
                             variant="outlined"
@@ -946,7 +1094,7 @@ ProductionLocationInfo.propTypes = {
     facilityProcessingTypeOptions: facilityProcessingTypeOptionsPropType,
     pendingModerationEventData: moderationEventsListItemPropType,
     pendingModerationEventFetching: bool,
-    pendingModerationEventError: array,
+    pendingModerationEventError: object,
     singleModerationEventItem: moderationEventsListItemPropType,
     singleModerationEventItemFetching: bool,
     singleModerationEventItemError: array,